backend/pyproject.toml
[tool.poetry]
name = "goklinik-backend"
version = "0.1.0"
description = "Go Klinik Clinical SaaS backend"
authors = ["Go Klinik Team <dev@goklinik.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.2.12"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.4.0"
django-cors-headers = "4.6.0"
drf-spectacular = "0.28.0"
python-dotenv = "1.0.1"
dj-database-url = "2.3.0"
psycopg2-binary = "2.9.10"
celery = "5.4.0"
redis = "5.2.1"
Pillow = "11.1.0"
django-storages = "1.14.4"
boto3 = "1.37.5"
supabase = "2.15.1"
gunicorn = "23.0.0"
firebase-admin = "6.6.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
